ボスのジャンプ攻撃解除を修正・ボスの壁攻撃後の方向修正・ 怯みカウントの修正・警告箇所の修正・ポーズ中のSEの一時停止を追加・ 一部コメントアウト箇所の削除

+++ DXBase/DXBase/sources/Actor/Person/Enemy/Bosses/Tornado.cpp
@@ -63,8 +63,7 @@ void Tornado::setAngle(int angle)
 	//	body_.GetBox().component_.point[i].x = tx * cosf(angle * MathHelper::Pi / 180) - ty * sinf(angle * MathHelper::Pi / 180) + position_.x;
 	//	body_.GetBox().component_.point[i].y = tx * sinf(angle * MathHelper::Pi / 180) + ty * cosf(angle * MathHelper::Pi / 180) + position_.y;
 	//}
-
-	angle_ = angle;
+	angle_ = (float)angle;
 }
 
 // ÉvÉåÉCÉÑÅ[ÇíTÇµÇ‹Ç∑
